requirements.txt
RPi.GPIO
appdirs==1.4.3
argcomplete==1.10.3
Pillow==6.2.1
pyserial==3.4
python-escpos==2.2.0
pyusb==1.0.2
PyYAML==5.2
qrcode==6.1
six==1.13.0
